Added Mapper class for parsing and result processing, fixed versioning
