+++ .gitignore
@@ -5,6 +5,9 @@
 #
 .github/
 
+#
+android/.settings/
+
 # XDE
 .expo/
 
@@ -67,3 +70,5 @@ lib/
 /package-lock.json
 /yarn.lock
 /example/ios/Podfile.lock
+android/java_pid19978.hprof
+android/java_pid20190.hprof
