+++ example/android/app/build.gradle
@@ -126,7 +126,7 @@ dependencies {
 
     // The version of react-native is set by the React Native Gradle Plugin
     implementation("com.facebook.react:react-android")
-    implementation 'com.monri:monri-android:3.1.+'
+    implementation 'com.monri:monri-android:3.1.1'
 
  
     if (hermesEnabled.toBoolean()) {
